requirements.txt
flask==2.3.3
elasticsearch==8.10.0
boto3==1.28.57
python-dotenv==1.0.0
pytest==7.4.2